requirements.txt
cx-Oracle==8.2.1
greenlet==1.1.2
python-dotenv==0.19.0
